changed from splash.js to index.js, touchups
