+++ pages/index.js
@@ -1,108 +1,139 @@
-import Head from 'next/head'
-import Image from 'next/image'
-import DateModified from '../components/atoms/DateModified'
-import styles from '../styles/Home.module.css'
-import { getLocalTopics } from './api/getData'
+import Link from 'next/link'
 
-export default function Home() {
+export default function Splash() {
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
-        <DateModified />
-
-        <p className={styles.description}>
-          Get started by editing{' '}
-          <code className={styles.code}>pages/index.js</code>
-        </p>
-
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/master/examples"
-            className={styles.card}
-          >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
+    <section className="flex h-screen">
+      <div
+        className="
+        z-0
+        fixed
+        bg-cover bg-center
+        inset-0
+        bg-gray-200
+        min-w-[18.75rem]
+        filter
+        blur-sm
+      "
+      ></div>
+      <div className="flex flex-col justify-center items-center m-auto">
+        <div className="z-10 bg-white h-auto w-[18.75rem] xl:w-[31.25rem]">
+          <img
+            className="h-auto w-64 mx-auto pt-6 xl:w-2/3 xl:mx-0 xl:px-6"
+            src="/sig-blk-en.svg"
+            alt="Government of Canada / Gouvernement du Canada logo"
+          />
+          <div className="flex w-max container py-11 mx-auto font-display">
+            <Link href="/en">
+              <a
+                className="
+                                      focus:ring-1 focus:ring-black focus:ring-offset-2
+                                      px-4
+                                      bg-blue-primary-deep
+                                      text-white
+                                      border border-blue-primary-deep
+                                      active:bg-blue-primary-deep
+                                      hover:bg-blue-bright
+                                      text-center text-md
+                                      h-11
+                                      xl:h-10
+                                      w-28
+                                      xl:w-36
+                                      py-3
+                                      mr-6
+                                      rounded
+                                      leading-3
+                                      shadow-md
+                                    "
+                id="english-button"
+                data-cy="english-button"
+                role="button"
+                draggable="false"
+                lang="en"
+              >
+                English
+              </a>
+            </Link>
+            <Link href="/fr">
+              <a
+                className="
+                                      focus:ring-1 focus:ring-black focus:ring-offset-2
+                                      px-4
+                                      bg-blue-primary-deep
+                                      text-white
+                                      border border-blue-primary-deep
+                                      active:bg-blue-primary-deep
+                                      hover:bg-blue-bright
+                                      text-center text-md
+                                      h-11
+                                      xl:h-10
+                                      w-28
+                                      xl:w-36
+                                      py-3
+                                      rounded
+                                      leading-3
+                                      shadow-md
+                                    "
+                id="french-button"
+                data-cy="french-button"
+                role="button"
+                draggable="false"
+                lang="fr"
+              >
+                Français
+              </a>
+            </Link>
+          </div>
         </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
+        <div
+          className="
+          relative
+          py-8
+          bg-gray-light
+          text-p
+          h-auto
+          min-w-[18.75rem]
+          w-[18.75rem]
+          flex
+          justify-between
+          p-6
+          xl:w-[31.25rem] xl:items-center
+        "
         >
-          Powered by{' '}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
-    </div>
+          <div className="w-28 text-base xl:text-p xl:w-max font-body text-blue-secondary-deep">
+            <Link href="https://www.canada.ca/en/transparency/terms.html">
+              <a
+                className="
+                                          inline-block
+                                          w-28
+                                          xl:w-max
+                                          mr-0
+                                          hover:underline
+                                          splash-a
+                                          text-lg
+                                        "
+                lang="en"
+              >
+                Terms &amp; conditions
+              </a>
+            </Link>
+            <span> • </span>
+            <Link href="https://www.canada.ca/fr/transparence/avis.html">
+              <a
+                className="
+                                          inline-block
+                                          hover:underline
+                                          font-body
+                                          text-lg
+                                        "
+                lang="fr"
+              >
+                Avis
+              </a>
+            </Link>
+          </div>
+          <img className="h-auto w-24 xl:w-28" src="/wmms-blk.svg" alt="" />
+        </div>
+      </div>
+    </section>
   )
-}
-
-export async function getStaticProps() {
-  let topicsData = []
-  let errorCode = false
-
-  //
-  // IF content enabled get the data from the api
-  //
-
-  if (!process.env.NEXT_CONTENT_API) {
-    // const { apiData, error } = await getTopics(locale);
-
-    let topics = []
-
-    // extract data from apiData then add it to the array topics
-
-    topicsData = topics
-    // errorCode = error;
-    errorCode = false
-  } else {
-    //
-    // Else get the content from the local file
-    //
-    const { localData } = getLocalTopics()
-
-    topicsData = localData
-    errorCode = false
-  }
-
-  return {
-    props: {
-      topicsData,
-      errorCode,
-    },
-  }
 }
